Created ServicesMonitor and refactored CollaborationMonitor
